refactor: Move all infrastructure packages to root directory infrastructure
